admin 폴더 구조 조정 -> 세부로 나눌수 있게 customer package 생성
